rag_chatbot_k8_test/vector-db/requirements.txt
langchain_community
langchain-openai
langchain
fastapi
pydantic
uvicorn
faiss-cpu
pickle